grpo run

+++ areal/core/remote_inf_engine.py
@@ -620,7 +620,8 @@ class RemoteInfEngine:
             # Update LoRA state if this was a LoRA update
             if meta.use_lora:
                 self.lora_initialized = True
-            shutil.rmtree(meta.path, ignore_errors=True)
+            if meta.clear_checkpoint:
+                shutil.rmtree(meta.path, ignore_errors=True)
 
         fut.add_done_callback(callback)
